ALF-3037: Не удаляются некоторые документы
git-svn-id: http://svn.datateh.ru/alfrescoecm/Business%20Platform/trunk@8364 697f8a5b-bb4b-4f4c-ab90-ff1ad96ece97

+++ lecm-documents/repo/src/main/java/ru/it/lecm/documents/policy/DocumentAttachmentsPolicy.java
@@ -116,7 +116,7 @@ public class DocumentAttachmentsPolicy extends BaseBean {
 				new JavaBehaviour(this, "onAddAttachment"));
 		policyComponent.bindAssociationBehaviour(NodeServicePolicies.OnDeleteAssociationPolicy.QNAME,
 				DocumentAttachmentsService.TYPE_CATEGORY, DocumentAttachmentsService.ASSOC_CATEGORY_ATTACHMENTS,
-				new JavaBehaviour(this, "onDeleteAttachmentAssoc"));
+				new JavaBehaviour(this, "onDeleteAttachmentAssoc", Behaviour.NotificationFrequency.TRANSACTION_COMMIT));
 	}
 
 	public void onCreateNode(ChildAssociationRef childAssocRef) {
@@ -308,9 +308,11 @@ public class DocumentAttachmentsPolicy extends BaseBean {
 		NodeRef attachment = nodeAssocRef.getTargetRef();
 		NodeRef category = nodeAssocRef.getSourceRef();
 
-		final NodeRef document = documentAttachmentsService.getDocumentByCategory(category);
-		if (document != null) {
-			onDeleteAttachment(attachment, document, this.documentAttachmentsService.getCategoryName(category));
+		if (nodeService.exists(category)) {
+			final NodeRef document = documentAttachmentsService.getDocumentByCategory(category);
+			if (document != null) {
+				onDeleteAttachment(attachment, document, this.documentAttachmentsService.getCategoryName(category));
+			}
 		}
 	}
